% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsthemes.R
\name{geom_label_mps}
\alias{geom_label_mps}
\title{MPS Standardized labeling function}
\usage{
geom_label_mps(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...,
  vjust = 0.5,
  fill = "white",
  size = 4,
  fontface = "bold",
  label.size = 1,
  label.r = unit(2.5, "pt")
)
}
\arguments{
\item{mapping}{A mapping aesthetic, assigned with aes()}

\item{data}{What data is being used? Inherited from initial ggplot() call}

\item{stat}{Default stat}

\item{position}{Default position}

\item{na.rm}{Whether to remove NA values from the labels. Default is FALSE.}

\item{show.legend}{Whether to show the legend from the labels. Default is FALSE}

\item{inherit.aes}{Whether to inherit aes from initial ggplot() call. Default is TRUE}

\item{...}{Other parameters to be passed to ggplot layer}

\item{vjust}{Default vertical justification}

\item{fill}{Default label background color}

\item{size}{Default text size}

\item{fontface}{Default face}

\item{label.size}{Default size of label}

\item{label.r}{Default size of label corner radius}
}
\description{
MPS Standardized labeling function
}
