% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpslatex.R
\name{mps_callout}
\alias{mps_callout}
\title{Callout Boxes, for PDFs}
\usage{
mps_callout(text, type = "note", title = NULL)
}
\arguments{
\item{text}{Whatever text you want included inside the callout box}

\item{type}{Type of callout box. Available choices are note, warning, important, and special.}

\item{title}{Only used when type = "special". This will create a custom title for the callout box.}
}
\value{
A LaTeX special color box function
}
\description{
Callout Boxes, for PDFs
}
