% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpshtml.R
\name{mps_callout_html}
\alias{mps_callout_html}
\title{Callout Boxes, for HTML docs}
\usage{
mps_callout_html(text, type = "note", title = NULL)
}
\arguments{
\item{text}{Whatever text you want included inside the callout box}

\item{type}{Type of callout box. Available choices are note, warning, important, and special.}

\item{title}{Only used when type = "special". This will create a custom title for the callout box.}
}
\value{
An HTML div function
}
\description{
Callout Boxes, for HTML docs
}
