% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpslatex.R
\name{mps_caption}
\alias{mps_caption}
\title{Add a LaTeX caption below a table, plot, or figure.}
\usage{
mps_caption(caption, type)
}
\arguments{
\item{caption}{Your caption}

\item{type}{One of "plots", "tables", and "figures"}
}
\description{
Function adds a caption using LaTeX font styling and automatically counts up
from 1 for each time the caption function is called.
}
