% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpshtml.R
\name{mps_caption_html}
\alias{mps_caption_html}
\title{Add a HTML caption below a table, plot, or figure.}
\usage{
mps_caption_html(caption, type)
}
\arguments{
\item{caption}{Your caption}

\item{type}{One of "plots", "tables", and "figures"}
}
\description{
Function adds a caption using HTML font styling and automatically counts up
from 1 for each time the caption function is called.
}
