% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpscolors.R
\name{mps_cols}
\alias{mps_cols}
\title{Function for retrieving hex codes of MPS official colors.}
\usage{
mps_cols(...)
}
\arguments{
\item{...}{Character names of MPS colors. Valid options include: yellow, blue,
red, teal, orange, wine, green, dblue, purple, gray, light gray, light yellow,
light blue, light red, light teal, light orange, light green, light dblue,
light wine, elight wine.}
}
\value{
A hex value for the MPS color
}
\description{
Function for retrieving hex codes of MPS official colors.
}
\examples{
# Return the hex value of MPS Wine
mps_cols("wine")

# Return the hex value of MPS red and MPS blue
mps_cols("red", "blue")

}
