% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsthemes.R
\name{mps_fonts}
\alias{mps_fonts}
\title{Force load the MPS Fonts if package doesn't automatically load them}
\usage{
mps_fonts()
}
\description{
A function to force load the Google Poppins font if not automatically loaded by require(mpstheme)
}
