% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpscolors.R
\name{mps_pal}
\alias{mps_pal}
\title{Function to interpolate an MPS color palette}
\usage{
mps_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in mps_palettes.}

\item{reverse}{Boolean to reverse the palette order}

\item{...}{Args to pass to colorRampPalette()}
}
\description{
Function to interpolate an MPS color palette
}
