% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpslatex.R
\name{mps_tcolor}
\alias{mps_tcolor}
\title{Change the font color of text in an R Markdown document output}
\usage{
mps_tcolor(color, text, mps = TRUE)
}
\arguments{
\item{color}{Color for the text. Available choices when mps = TRUE are:
yellow, blue, red, teal, orange, wine, green, darkblue, purple, and gray.
Available choices when mps = FALSE are: black, blue, brown, cyan, darkgray, gray,
green, lightgray, lime, magenta, olive, orange, pink, purple, red, teal,
violet, white, yellow.}

\item{text}{Whatever text you want to change the color of.}

\item{mps}{Boolean for whether you want the MPS official colors \link{TRUE}, or one of the
default LaTeX colors \link{FALSE}.}
}
\value{
A LaTeX textcolor function
}
\description{
Change the font color of text in an R Markdown document output
}
