% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpshtml.R
\name{mps_tcolor_html}
\alias{mps_tcolor_html}
\title{Change the font color of text in an R Markdown document HTML output}
\usage{
mps_tcolor_html(color, text)
}
\arguments{
\item{color}{Color for the text. Available choices when mps = TRUE are:
yellow, blue, red, teal, orange, wine, green, purple, and gray.}

\item{text}{Whatever text you want to change the color of.
#'}
}
\value{
A HTML span color function
}
\description{
Change the font color of text in an R Markdown document HTML output
}
