% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpscolors.R
\name{scale_color_mps}
\alias{scale_color_mps}
\title{Color Scale of MPS Colors for ggplot2}
\usage{
scale_color_mps(palette = "standard", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of an MPS Palette. Valid options include:
primary, secondary, dark, light, hot, warm, mild, frigid, cold, cool, all, and colorful.}

\item{discrete}{Boolean for whether aesthetic is discrete. Default is TRUE.}

\item{reverse}{Boolean for whether to reverse palette. Default is FALSE.}

\item{...}{Additional args passed to discrete_scale() or scale_color_gradientn()}
}
\description{
This function allows a user to set a color scale based on MPS colors.
}
