% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsthemes.R
\name{theme_ofda_donut}
\alias{theme_ofda_donut}
\title{MPS Theme for ggplot2 donut plots - for OFDA department}
\usage{
theme_ofda_donut(
  pri.color = mps_cols("wine"),
  sec.color = mps_cols("light wine"),
  ter.color = "gray40",
  ...
)
}
\arguments{
\item{pri.color}{Primary color of the theme. Impacts title, subtitle, legend title, and facet panels.}

\item{sec.color}{Secondary color of the theme. Impacts axis titles and legend title.}

\item{ter.color}{Tertiary color of the theme. Impacts axis labels, captions, and legend text.}

\item{...}{Other arguments to be passed to theme()}
}
\description{
Basic theme for MPS graphs. Based on Fiona Adams' excellent School Transformation Theme.
Primary, Secondary, and Tertiary colors can be chosen from any color palette, or from mps_cols().
}
