% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsthemes.R
\name{theme_rea}
\alias{theme_rea}
\title{MPS Theme for ggplot2 - for REA department}
\usage{
theme_rea(
  pri.color = mps_cols("red"),
  sec.color = mps_cols("light red"),
  ter.color = mps_cols("gray")
)
}
\arguments{
\item{pri.color}{Primary color of the theme. Impacts title, subtitle, legend title, and facet panels.}

\item{sec.color}{Secondary color of the theme. Impacts axis titles.}

\item{ter.color}{Tertiary color of the theme. Impacts axis labels, captions, and legend text.}
}
\description{
Basic theme for MPS graphs. Based on Fiona Adams' excellent School Transformation Theme.
Primary, Secondary, and Tertiary colors can be chosen from any color palette, or from mps_cols().
}
