% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsthemes.R
\name{title_mps}
\alias{title_mps}
\title{MPS Report Title - for OFDA department}
\usage{
title_mps(title = "Some Title", color = mps_cols("wine"))
}
\arguments{
\item{title}{The report title.}

\item{color}{Background color of the title box.}
}
\description{
Function to insert MPS-styled report titles.
}
