% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpslatex.R
\name{increment}
\alias{increment}
\title{Simple incremental function to create new function...}
\usage{
increment(init)
}
\arguments{
\item{init}{Initial value, default should be 0}
}
\description{
Simple incremental function to create new function...
}
